requirements.txt
pyqt5
pylnk3
pandas
pyshark
striprtf
olefile
Pillow
scandir
pefile